requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.5
python-multipart==0.0.9
pydantic==2.8.2
starlette==0.38.2
numpy==1.26.4
sentence-transformers==2.7.0
scikit-learn==1.4.2
tiktoken==0.7.0
httpx==0.27.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
